Move Backup Discovery to azure.go
