+++ service/discovery/azure/azure_test.go
@@ -36,10 +36,13 @@ import (
 	"testing"
 
 	"clouditor.io/clouditor/internal/testdata"
+	"clouditor.io/clouditor/internal/util"
+	"clouditor.io/clouditor/voc"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/dataprotection/armdataprotection"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/storage/armstorage"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/subscription/armsubscription"
 	"github.com/stretchr/testify/assert"
@@ -325,7 +328,8 @@ func NewMockAzureDiscovery(transport policy.Transporter, opts ...DiscoveryOption
 				Transport: transport,
 			},
 		},
-		csID: testdata.MockCloudServiceID,
+		csID:      testdata.MockCloudServiceID,
+		backupMap: make(map[string]*voc.Backup),
 	}
 
 	// Apply options
@@ -334,4 +338,117 @@ func NewMockAzureDiscovery(transport policy.Transporter, opts ...DiscoveryOption
 	}
 
 	return d
+}
+
+func Test_azureStorageDiscovery_handleBackupVaults(t *testing.T) {
+	type fields struct {
+		azureDiscovery     *azureDiscovery
+		defenderProperties map[string]*defenderProperties
+		backupMap          map[string]*voc.Backup
+	}
+	type args struct {
+		vaults []*armdataprotection.BackupVaultResource
+	}
+	tests := []struct {
+		name   string
+		fields fields
+		args   args
+		want   assert.ValueAssertionFunc
+	}{
+		{
+			name: "Empty input",
+			fields: fields{
+				azureDiscovery:     NewMockAzureDiscovery(newMockStorageSender()),
+				backupMap:          make(map[string]*voc.Backup),
+				defenderProperties: make(map[string]*defenderProperties),
+			},
+			args: args{
+				vaults: []*armdataprotection.BackupVaultResource{},
+			},
+			want: func(tt assert.TestingT, i1 interface{}, i2 ...interface{}) bool {
+				s, ok := i1.(*azureStorageDiscovery)
+				if !assert.True(tt, ok) {
+					return false
+				}
+
+				return assert.Equal(t, 0, len(s.backupMap))
+			},
+		},
+		{
+			name: "Error getting instances",
+			fields: fields{
+				azureDiscovery:     NewMockAzureDiscovery(newMockNetworkSender()),
+				backupMap:          make(map[string]*voc.Backup),
+				defenderProperties: make(map[string]*defenderProperties),
+			},
+			args: args{
+				vaults: []*armdataprotection.BackupVaultResource{
+					{
+						ID:       util.Ref("/subscriptions/00000000-0000-0000-0000-000000000000/resourcegroups/res1/providers/Microsoft.DataProtection/backupVaults/accountDoesNotExist"),
+						Name:     util.Ref("accountDoesNotExist"),
+						Location: util.Ref("westeurope"),
+					},
+				},
+			},
+			want: func(tt assert.TestingT, i1 interface{}, i2 ...interface{}) bool {
+				s, ok := i1.(*azureStorageDiscovery)
+				if !assert.True(tt, ok) {
+					return false
+				}
+
+				return assert.Equal(t, 0, len(s.backupMap))
+			},
+		},
+		{
+			name: "Happy path",
+			fields: fields{
+				azureDiscovery:     NewMockAzureDiscovery(newMockStorageSender()),
+				backupMap:          make(map[string]*voc.Backup),
+				defenderProperties: make(map[string]*defenderProperties),
+			},
+			args: args{
+				vaults: []*armdataprotection.BackupVaultResource{
+					{
+						ID:       util.Ref("/subscriptions/00000000-0000-0000-0000-000000000000/resourcegroups/res1/providers/Microsoft.DataProtection/backupVaults/account1"),
+						Name:     util.Ref("account1"),
+						Location: util.Ref("westeurope"),
+					},
+				},
+			},
+			want: func(tt assert.TestingT, i1 interface{}, i2 ...interface{}) bool {
+				s, ok := i1.(*azureStorageDiscovery)
+				if !assert.True(tt, ok) {
+					return false
+				}
+
+				want := &voc.Backup{
+					RetentionPeriod: 0,
+					Enabled:         true,
+					GeoLocation:     voc.GeoLocation{Region: "westeurope"},
+					Storage:         voc.ResourceID("/subscriptions/00000000-0000-0000-0000-000000000000/resourcegroups/res1/providers/Microsoft.DataProtection/backupVaults/account1/backupInstances/container1-container1-22222222-2222-2222-2222-222222222222"),
+					Policy:          "policyId",
+				}
+
+				val, ok := s.backupMap["/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Storage/storageAccounts/account1"]
+				assert.True(t, ok)
+
+				return assert.Equal(t, want, val)
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			d := &azureStorageDiscovery{
+				azureDiscovery:     tt.fields.azureDiscovery,
+				defenderProperties: tt.fields.defenderProperties,
+			}
+
+			// initialize backup instances client
+			_ = d.initBackupInstancesClient()
+
+			d.handleBackupVaults(tt.args.vaults)
+
+			tt.want(t, d)
+		})
+	}
 }
