+++ service/discovery/azure/network.go
@@ -44,6 +44,7 @@ func NewAzureNetworkDiscovery(opts ...DiscoveryOption) discovery.Discoverer {
 		&azureDiscovery{
 			discovererComponent: NetworkComponent,
 			csID:                discovery.DefaultCloudServiceID,
+			backupMap:           make(map[string]*voc.Backup),
 		},
 	}
